requirements.txt
Flask
google-genai 
python-dotenv 
PyPDF2 
python-docx